Made the SlideShow a component and added component DarkModeProvider, which acts as a context provider for dark/light mode, making it simpler to switch between modes. Also added a type for each project.
