+++ src/types/project.ts
@@ -1,0 +1,8 @@
+export interface Project {
+  id: number;
+  title: string;
+  description: string;
+  image: string;
+  githubUrl: string;
+  technologies: string[];
+}
